Remove User Agent Starboard Stability Suffix
Per discussion in go/starboard-notes meeting May 2.

Currently, when Cobalt is compiled with a version of Starboard
marked "experimental" or "release candidate", this is placed in the
user agent string as a way to catch such non-ideal configuration at
cert time.

However, the new proposed branching scheme's "update number" combined
with the other fields will serve the same function.

So, remove this field from the user agent.

Change-Id: I7a568a53ea54d277bb1f51af283b08001a7842ad

Former-commit-id: d8fa1026a89a3932f942af969b75e530ab5727cd