Fixes running --url on any xbox target.
Missing dependency framework error now detected properly
and the tester/developer is now warned for corrective action.

NetArgs and NetLog thread on the launcher side are now started
before the target binary is executed. This fixes an issue
with Xbox thinking that the app hasn't launched due to socket
wait for args/log connection.

Auto sign in now implemented. The launcher will now check
the xbox to see if anyone is signed in, if not then any
users that exist will be automatically signed in. If no users
exist on the xbox then the launcher will fail immediatly with
instructions for the user to create an account on the device.

Lots of additional API implementations and command line tests
for the in the xbox network layer.

Change-Id: I8cbc739fad21b0391a27bc72d9b27622e9010a2c

Former-commit-id: 26545ebc17f85615bc74292368a2b0fd4c9cc8b3