Disable distance field fonts when deterministic rendering is needed.
This allows Scuba screen diff tests to pass without being flaky.

Presumably the reason for this is that the glyph atlas will be
populated with the size of the first glyph seen, and this glyph
will be re-used for similarly sized glyphs, so for pixel-perfect
determinism, we would have to make sure that glyphs enter the
atlas in the same sequence, and Cobalt can't make that guarantee.

Since what we're really looking for here is deterministic
rendering, while still available, the
'--disable_rasterizer_caching' flag is renamed to the more general
'--force_deterministic_rendering'.  Currently they are synonyms
for backwards compatibility, but eventually it would be nice
to deprecate 'disable_rasterizer_caching' in favor of the more
descriptive 'force_deterministic_rendering'.

b/78363319

Change-Id: Ibe6ea7b978caa5e00eaa8720bbdd8f535c2d4212

Former-commit-id: d5bd561170d6108e4ec7bd6032fbaf206d033f9c