Fix JavaScript engine standalone shells
Fix a compilation error in the JavaScript engine standalone shells.  The
break got checked in because the shells do not currently get compiled as
part of all.

In order to ensure that this doesn't happen again in the future, make
it so they *do* get built as part of all.

Change-Id: Iaa335100fb42d403ee98de39bab0549892d7d0ec

Former-commit-id: 412552c0e79002851b8b8ed5ab67c7e0336dd43e