Keep mouse/pointer event ordering relative to key events.
As an optimization before, we were batching mouse and pointer
events and processing them all at once, at layout time.  This
allowed the document to complete the layout it would need to do
to properly process the events, as well as not trigger any
intermediate layouts.

However it does mean that key events, which are not queued, do not
get received by the app JavaScript in the correct order relative
to mouse events.

This change makes it so that whenever there is a non-mouse/pointer
input event, it first stops and processes all pending mouse/pointer
events, ensuring the correct order.

In terms of performance implications, it could mean that we do an
extra layout now if a key is pressed while there are pending pointer
events, however this would not only be rare, but it also can
be considerably (or completely) mitigated by partial layout support.

b/72493981

Change-Id: Id41e8b385a9723bbb07ad5b8d3cf7c91a96f53d9

Former-commit-id: 1cae29d2be49dfbdbac1466d79659f01b246c144