Stop tracking first frame in rasterize interval stats
Previously |rasterize_periodic_interval_timer_| and
|rasterize_animations_interval_timer_| both included the first duration
of the first rasterization in their stats. However, this value is
different than the rest of their values, because it does not include a
full interval. Given that it doesn't match up with the rest of the
values and confuses the results, it is no longer being included.

However, it is still included in |rasterize_animations_timer_|, which
only tracks the rasterization duration and not the full interval.

Change-Id: I3a214ba583de0983cf145f9981dd9c33f464516f

Former-commit-id: aed9733d02f196d6f88a2661fdbbacf874119890