Initialize document ready state to complete if there is no browsing context
Previously, if |HasBrowsingContext| was false in |Document|'s
constructor, the |ready_state_| field would not be initialized at all,
which would potentially go on to cause an error in bindings when
attempting to convert |ready_state_| into a JavaScript enum value.

This change addresses the issue by initializing |ready_state_| to
|kDocumentReadyStateComplete| if no browsing context is available, in a
manner similar to what Chromium does.

See:
https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/dom/document.cc?l=614&rcl=967adc730ab3b738d8d8f65a774dcf3a5061cd47

b/69181029
b/77937131

Change-Id: I3013a82da739a1570e53e0dd6370090822744e3d

Former-commit-id: 8ab47261a6d6be1ec39968f9c2ee3a4242d26877