Fix the window.onerror handler not expanding event arguments.
The special case that is addressed here is when setting the
event handler for window.onerror.  Not window.addEventListener(),
and not non_window_event_target.onerror, only window.onerror.

In order to deal with this special case in as clean a way as
possible, a new GenericEventHandlerReference class is created to
abstract the difference between EventListener and the new
OnErrorEventListener.

Some other code such as ScriptEventLog has been factored out
of EventListener into its own .h and .cc file.

b/73776662

Change-Id: I443ffd625358338135c5336083fc99db557691ad

Former-commit-id: 6967af9ec501ce1a1802809e10aa39aa651e5709