Ensure platform_path is absolute when calling SetDirectory().
In most cases, platform_path would be the output of
starboard.tools.platform._EnumeratePlatforms() which would
produce absolute paths.  However, when calling the app launcher
scripts from x20, platform.py is instead generated by
starboard.tools.app_launcher_packager.WritePlatformsInfo() which
explicitly saves the paths relative to the repository root.

So basically, before this CL, SetDirectory() would usually
be called with an absolute directory when called from within
the Cobalt codebase, but called with a relative directory when
called from x20 (and this would result in errors).  After
this CL, it's always called with an absolute directory.

b/80301701

Change-Id: I75a048d6d93c6b2a7b6655d971fa8f729662569f

Former-commit-id: 1233e531ad89cd8b2f28ef7c3ed461a8887017ab