Override OnResume() to tell StarboardBridge when resuming.
This ensures we call StarboardBridge.beforeStartOrResume() when resuming
for any reason. Previously, we wouldn't call that when resuming and the
window didn't get destroyed/re-created, as is the case when Daydream
starts. That caused MediaSession (and FeedbackService) to remain idle
after coming out of Daydream, so we wouldn't request a wakelock for
subsequent video playbacks.

b/80259656

Change-Id: Ifdbf517111d518077567a55a63e9a8f4fc9797d5

Former-commit-id: 7eb898215136d67f453eeacb35b7a5e577d68a00