Make body background clear to transparent instead of not drawing.
Formerly, Cobalt would interpret a "background-color: transparent"
on the initial containing block the same way it would for any
other block; with blending enabled.  This implies that no background
would be rendered for initial containing blocks with a transparent
background, which is the default background color.  This change
makes it so by default, this will instead result in the a draw
call with blending disabled, effectively clearing the screen.

This makes Cobalt's behavior much more intuitive, and eliminates
the unusual "hall of mirror" effects when running Cobalt on
web pages that don't fill their backgrounds with content (like
layout tests).

Additionally, the EGL rasterizer will no longer clear the screen
by default, as it was doing before.  It now respects the
kSubmitFlags_Clear flag.

b/22985529

Change-Id: Icbfb57e18dbfaf4148458fd5948676fe62d25a0b

Former-commit-id: 90eda90f1d691597dbd5fc593de8c6223e1ad62c