Disable synthetic bolding with NotoColorEmoji.ttf
Synthetic bolding doesn't work with color emojis and causes the glyphs
to become invisible. Now that Airstream is using bold titles, this was
exposed as an issue. Disabling synthetic bolding causing the color
emojis to again be visible.

Updated the color emoji layout test to test this case.

This fixes b/110282063.

Change-Id: If179d2026d18531c9208b603031336ca704a0bcb

Former-commit-id: b8e08d3223572d07382f4718bcffbd3acd0ba3b1