Tolerate directfb 1.7.7 and older shutdown crash
It turns out that directfb 1.7.7 has an uninitialized variable
that causes crashes on teardown:

http://lists.openembedded.org/pipermail/openembedded-core/2016-June/122843.html

We've also seen problemx on 1.2.x.

Since we currently get directfb as a Rodete/dpkg for our distributions,
the easiest thing to do is simply work arount it by catching the SIGSEGV
and longjmp'ing past the crash.

We also call _exit() in this case so as to avoid problematic atexit handlers

Note this will result in shutdown memory leaks, but at least there
shouldn't be flakes.

b/74449762

Change-Id: Idba774a18308fdacd6498254efbc53c4f60695bd

Former-commit-id: bded474ab70aedfec42b584041c277adfc2ed79a