Fix V8 not passing Window.onerror Qual-E test.
V8 was passing the opposite boolean value for the muted
variable into its script origin object.  This means that
we were muting JavaScript errors when we weren't supposed
to, and not muting them when we were.

Additionally, html_script_element.cc was deciding that
any script with a file:// scheme should always be muted,
because file:// would formerly end up with a opaque
origin which would always equate to false with other
opaque origins.  This logic is now adjusted so that all
file:// scheme URLs share the same origin.

The previous paragraph implies that our layout tests have
been wrong for quite some time, in that they were reporting
errors (for say, the window-onerror.html test) when they weren't
supposed to.  V8 is a new engine so it can be forgiven, but
SpiderMonkey seems to have always been ignoring the mute
request, so this is patched now also.

b/110484264

Change-Id: Iea708518655bce4854e94800fb92bc89aa964f1c

Former-commit-id: f2b742e3a127b9b38ad32e4c5bc5e20e843d1f47