Notify ready state ended on stopped tracks
Previously, one could add a sink to a stopped track, and it would
never find out that the track had never been started, or is already
ended.

Now, sinks are only added to non-stopped tracks. If an attempt
to add a sink to a stopped track is made, the sink will be notified
of the ended ready state.

This matches Chromium behavior.

b/79546209

Change-Id: Ib99556a2488f8988ab7f44c5b0631e2470fadbdf

Former-commit-id: 199f60bf64eef0d3f0053d6d11868320d9f24caf