Add and use an interface for audio encoding
This CL adds an interface (AudioEncoder) that is used by
|media_capture::MediaRecorder|.  This can be extended to add more
encoders like FLAC in the future.

Also, the |MediaRecorder| constructor throws a |DOMException| if an
invalid MIME type is passed in.

Additionally, there was a threading bug that was fixed by the refactor,
and a new test has been added to prevent it from happening again.

b/79545424

Change-Id: I28f304510f22eb394aa2e99f09a5e1e2c9d1ecbe

Former-commit-id: ffc76253529e29cbdebba0f4bbbc80af368861f5