Make material design spinner render without aliasing.
The Skia rasterizer was previously setup to avoid applying rotations
to offscreen renders, the reasoning being that if these were applied
when blitting the final unrotated offscreen surface, it would
save surface space.  However as it turns out, doing this can cause
aliasing since the offscreen rendering does not know that it will
eventually be rotated.

Thus, we instead we just apply these rotations to the offscreen
surface rendering, and accept that it may increase the size of
the offscreen surfaces.

This is a change to the skia rasterizer, however all other
rasterizers will fall back to this logic for the provided
material design spinner demo.

b/67861162

Change-Id: If95350f419c4b45adafd49945928ef2b3e6b5483

Former-commit-id: c07d9a01c3269518338ccc3a8e8394a8f0ce8be2