Revert "Revert "Modify format of version string to newly proposed format.""
This reverts commit 09281d762e49eec3baf58a9740c2fcad49595d2e.

This will be submitted after
https://cobalt-review.googlesource.com/c/third_party/starboard/roku/+/4590
is submitted to ensure that Roku builds do not break.

Also, I have updated the version to read 19 as it should for launch
instead of 18 which...  really didn't make sense, I guess I just
was nervous about pulling the trigger on 19.  But we are cutting
very soon.

Change-Id: Ib6faeb470c786c00a19a727c125dbc4d08ffa621

Former-commit-id: 52417af7095fcd9fb35e40da4537fdee6b1561ef