Add back nplb microphone tests on Android.
Check for a connected microphone in
SbMicrophoneImpl::GetAvailableMicrophones() to prevent nplb tests from
failing when no microphone is connected.

While the nplb tests use SbMicrophoneImpl::GetAvailableMicrophones() and
check if a microphone is connected, SbMicrophoneImpl::Open() now checks
if a microphone is disconnected (rather than connected). This
distinction is necessary because there is no way of checking for a
microphone/device before API 23.

Intended behavior for APIs < 23:
*SbMicrophoneImpl::Open() should assume that a microphone is connected.
If that is not the case, then the implementation will later return false
when there is an error realizing the audio recorder.
*The nplb tests (and SbMicrophoneImpl::GetAvailableMicrophones()) should
assume that no microphone is connected because the tests will fail if
they are run under the false initial assumption that a microphone is
connected.

b/110433932

Change-Id: I76e4edcf9122e9fad5adaa3297b114d2fafcbb6f

Former-commit-id: f2f3eeed0269c5a657f6f6969f9b3775ab4e4154