Move android compiler flags into its abstract toolchain
Allow the platforms using the clang toolchain to specify defines that
will be merged with the defines in the gyp files.

Pruned warning flags to only disable those that are necessarry for
Android builds. Note that -Wconversion is NOT set as there are many
implicit type conversions all over the code base.

Change-Id: I54156c8c4c558097ef4652ddb599335c79d7003c

Former-commit-id: d1d6458e236d4e4e67807a2052af722fa06aa0f3