Make ~Document() tolerate page_visibility_state_ not existing.
This change relaxes ~Document()'s requirement that
page_visibility_state_ exist.  This is implemented by making
page_visibility_state_ a weak pointer.  This is done because
it is possible for the Window object, which owns
HTMLElementContext, which owns PageVisibilityState, to be
destroyed before Document.

Essentially, this is a result of JavaScript garbage collection
not being deterministic.

b/110953238

Change-Id: Icdb389061c856c9788e5692a331c19fe50be5acc
(cherry picked from commit 9feaf4570a4c75804431e73eb3d207d518cc93ee)

Former-commit-id: 174deef6c80e87c065eab6dc9f41e275a4de59d6