Fix ninja error of isolate_fellowship.cc after clean
To touch isolate_fellowship.cc and rebuild it everytime v8 code
changes, we put it as an output in gyp. However, when running
ninja clean, ninja deletes all output file which means we should
never put cobalt code in any output list in gyp.
I'm adding a dummy target as the output. This output will not be
generated but just serve to trick gyp.

Change-Id: I5ba7b3f8593f791966d7acb7ac55aa47ec35eef0

Former-commit-id: 7d4d374387c71d509e635fcead6596b5c2b18cea