Improve logging consistency.
This removes the tickcount (monotonic time) from the log message, and adds
the date and time of day with sub-second precision.

This makes the timestamps from SB_LOG()/SB_DLOG() match those of LOG()/DLOG().

Change SB_DLOG() into SbLog() for the messages printed in
MemoryReportingTest, since now there can be allocations from the ICU
initialization, and there will be allocations from the timezone
determination inside ICU.

This moves the eztime dependency to the starboard target to ensure that all
starboard targets have the required functions.

Change-Id: Ibe02f3490734bc1e5c93e9fc08be728ad0a8fc81

Former-commit-id: 86be9b043d925a29b74ed059ebf030123511787a