Ensure window.setTimeout() doesn't return 0 when shutting down.
Originally we would destroy our entire WindowTimers framework,
including the unique ID management aspect of it, when we're about
to shutdown the WebModule.  However, JavaScript may still be
running at this time and may call setTimeout(), in which case
0 each time resulting in spam and potentially incorrect behavior
in that the ID returned is no longer unique, as it is expected
to be.

This fixes that by keeping the WindowTimers object alive, but
just disabling its callbacks from being called.

b/76117263

Change-Id: I5f529d85362427726a80e46aba49e46704534a6d

Former-commit-id: 1d1a21ff4005fbe2ec6590e431b1f6e915022bd6