Delete non-specialized storage_write_record.cc
Recently, a POSIX and Win32 specialization was added for
storage_write_record.cc.

All (non-contrib) use cases use either one of these two specializations
or their own custom implementations.

So, we'll delete this one since as it stands it is subject to b/78941867
and we think no one should be using it.

Note we do not plan on taking this change for the 19.lts branch because
we'd like to not disrupt partners, and b/78941867 is not critical.

Change-Id: I47a18a6cfd432fdfa728675c273155d933dbfee6

Former-commit-id: c2caef3daf51a66ace27c9247fe4032ed488e7b0