Use less sockets in SbSocketCreateTest.ManyTcpAtOnce
The test used to create 128 tcp sockets, which fails on Switch as it
only allows for 32 concurrent tcp sockets.
Reduce the number of sockets to create in the test to 32.  If Cobalt
ends up not be able to run properly using 32 tcp sockets then we will
probably have to refactor Cobalt.

b/113124009

Change-Id: Ic100453a880c9b3e03c65a26c696e0acfd4514ef

Former-commit-id: 89ef37342054ecbe2340cef342a237f0e72e0500