Set position and speed on the Android MediaSession.
This allows Assistant to do relative seeks since it's now able to get
the accurate playback position from its MediaController.

The SbPlayer is short-circuited from the Android player implementation
to the Android MediaSessionClient implementation. This allows it to
learn the position and speed with SbPlayerGetInfo2 in lieu of the
missing information not available through the web MediaSession.

Note that duration is NOT included since it's not ever set by the
shared Starboard player internal implementation. Setting it to the wrong
value negatively affects Assistant, so we don't set it at all.

This should not be a permanent fix, as the web MediaSession should be
extended to allow live playback metadata to be communicated from the web
app to the user agent.
https://github.com/WICG/mediasession/issues/70
https://github.com/WICG/mediasession/issues/202

b/111302299

Cherry-Pick: I9e29176b43f06473fdb6a990dfd088908aa2fa9f
Change-Id: I099004d086b2b85420551a0d5c83add2877f310c

Former-commit-id: 52c6aba6ad456c16c94e28fcd778785ceda8c09d