Remove |partial_layout| debug command & refactor command line switch.
Move the processing of the command line switch out of web module and
into browser module so that there is no dependency on the definition of
browser switches in web module. The switch now only affects the main web
module directly owned by browser module (i.e. not splash screen nor
debug console).

Change the implementation to be a boolean (instead of the "on" or "off"
command strings) and process it as part of web module options during
initialization. As a boolean switch with a default value of |true|, it
has been renamed to |--disable_partial_layout|.

b/114104889

Change-Id: I12076ae50a5090d576358cf91f4e373e863c5bcd

Former-commit-id: fbf332325982162be869da90a1ba8b5fe7e71bd1