Revert "Change CommandHandlerMap from map to multimap."
We no longer have the |debug.partial_layout()| command, so we no longer
need the workaround of broadcasting through a multimap.  It's cleaner to
have a single registered command handler rather than treating it like a
broadcast to multiple handlers for the same command.

Cleanup of console commands, follow-on to b/31266139

This reverts commit 2852e3ee60cb366bd65d76ff3ab74cee10daa231.

Change-Id: I702df491e81f8b736e21da7a5ee7347b4d3ae83f

Former-commit-id: f52f8ff504172f3e36b33e44ffe6d1039667ebb4