Delete data/websocket/OWNERS
This OWNERS file came over from chromium and is not applicable
to the Cobalt sourcebase.

However, it ends up being copied into the content directory, and then
in turn imported into ***REMOVED*** when we update ***REMOVED*** binaries.

Since this OWNERS file is invalid for ***REMOVED*** (it's valid for
chromium), this causes errors when updating ***REMOVED***.

Change-Id: I27c99c26b7bae10e998afafb2d1b35d3e96f22a6

Former-commit-id: 1ffe87fc832ff22abe9c233836527ecaed8ffd28