Use musl on clang and x64
For targets already building with clang on an x64, use the portion
of musl that we've carved out.

The immediate impact is that we will use these math functions instead
of, in many cases, using the system ones.

The api_leak_detector whitelist shrinks as a result.

stripped linux-x64x11_gold/cobalt goes up 29k (out of 33.7m)

Change-Id: I595f96ef7f9c42a4bcfd9ed721e4a9389d9829fe

Former-commit-id: a2950768816308b0a9ddc516a51226ccecb3d13c