Update all the Android SDK components to latest versions.
This is a cascade of dependecies:
- NDK 17 => 18 - automatic update, doesn't include mips support.
- Gradle plugin 3.0.x => 3.1.4 - removes assumption about mips support
- Gradle itself 4.1 => 4.4 - required by Gradle plugin 3.1.4

Gradle plugin 3.1.4 doesn't support both "ndk {}" and "splits.abi {}" to
be specified, so removed the latter from app/build.gradle.

While at it, updated all the other components:
- Support libraries 27.1.0 => 27.1.1
- GMS auth 11.8.0 => 15.0.1 (not the latest, but works with other dependecies)
- Build tools 27.0.3 => 28.0.2
- lldb 3.0 => 3.1
- platform SDK 27 => 28 (did not update |compileSdkVersion| to 28)

Added platform-tools as a convienence to developers who will want it in
the SDK downloaded by cobalt.

b/116049435

Change-Id: Ic2215085ebde77ce0414067f2ab6adf8a3cac00d

Former-commit-id: b972b4bff7f2ad92ca55be645d950011bc64c38d