Fix SbTimeGetNow() resolution and time zone values.
The OS only gives a posix time with one second resolution. This uses the
system tick as an offset to return microsecond resolution for the time.

This also fixes the time zone values to give the correct current time zone
name and offset (previously, would return UTC and offset zero).

Adds checks to the corresponding time zone tests.

Change-Id: I41224b6e181ee697ac399260eef5ce9f363d8d45

Former-commit-id: 9c082ce039445c32dc0964b99aa1346fc8cbbd55