Fix black box tests' python error and port access issue
When running black box tests on devices, there are two current issues:
1. We create a new launcher to query if platform supports sending
signals and launcher needs device_id as well which we did not provide.
2. Black box test server sometimes uses a very high port number and
chosen port may not be available for remote access. So I changed it to
try port 8000 initially and if not available pick one from pool.

b/113043717

Change-Id: Ib9fb447f89c36130b6101bcf96f6943bdfcfc3b2

Former-commit-id: 33ff8d2443d3bbcd3cb14ed45aab10c2570cd83c