User selection, application state, and network error improvements.
This moves the user selection to the manifest, letting the system handle it.

This handles network errors by verifying if the network interface manager
reports that the network is available, and if (re)submit network requests,
and verify the network service account until it is reported available.

This adds a hook to verify network interface availability when socket
connections fail.

This enables system events that allow Cobalt to respond, suspend, resume,
and exit events, and injects the corresponding events.

b/117119243
b/117117415
b/117091924
b/117106652
b/112424653

Change-Id: Ic14ad3d353262ae1c692ec5abbdbe4bbfba1505a

Former-commit-id: 89827b87606d446bdf69394ded2005be8760bbec