Lay the groundwork to integrate V8 inspector
This puts the pieces in place to hook debugging protocol commands to V8
through a |V8Inspector|, but doesn't actually wire it up to be used yet.

Change the |ScriptDebugger| interface to allow directly sending protocol
messages to the JavaScript engine. Likewise, change its |Delegate|
interface to receive protocol responses and events, as well as letting
the engine control when we pause/resume the event loop for JavaScript
execution.

Implement a new |V8cScriptDebugger| that makes a single long-lived
connection to a |V8Inspector|. Also add a new |ScriptDebuggerComponent|
to send protocol commands and events to/from the |V8cScriptDebugger|.

The |JavaScriptDebuggerComponent| remains for now but doesn't really do
anything since it's based on a |ScriptDebugger| that's not implemented
for any current engine. However, its unused implementation of the
(changed) |ScriptDebugger::Delegate| interface has been severed.

The |RuntimeComponent| remains as well, and continues to work in its
limited functionality with both MozJs and V8 to support the on-screen
console.

Changed the on-screen console to initialize the executionContext id to
|undefined| so that prior to receving an |executionContextCreated| event
it will use the default context rather than specifying an invalid
context ID of 0.

Fixed the on-screen console to show the |value| of the response to an
|evaluate| command if it doesn't have a |preview| or |description|.

Fixed |DebugDispatcher::SendEvent| to accommodate null event parameters.

b/116317473

Change-Id: I5b6f8f35efc43d92075e4841c08b4bb61f604870

Former-commit-id: 9726c868427b7a6e3113d299eb2d9844f31c5e41