Register protocol domains rather than individual commands.
Components now register a single handler function for an entire domain
of protocol commands.

The |DebugDispatcher| calls the component's registered callback for all
commands within a domain, and if that returns an empty result it falls
back to trying to run a JavaScript implementation that the component
previously loaded. This simplifies hybrid components (DOM and Runtime)
since simply defining their JavaScript command functions is sufficient
without also having to wire them up individually in C++.

For those commands implemented in C++, the components all use the
|CommandMap| template as a map of pointers to their member functions
that handle individual commands. The template provides a |RunCommand|
to bind as the component's domain |CommandHandler|.

b/116317473

Change-Id: I5b9e6a1c5fd0be92735f23699ac4a08685cfb359

Former-commit-id: 8089faad8e7de9c0ee125705a06c079fe4832d16