Hooks up new starboard diagonal feature to trigger DPI change
See related cl in starboard:
https://lbshell-internal-review.git.corp.google.com/c/starboard/+/117400

This CL allows a change in the physical size of the screen,
expressed as the diagonal-length-inches, to propagate
to the media-query concept in CSS and javascript. See demo
cobalt/.../media-query-dpi-test.html

The first platform which needs this, nxswitch, will generate
a SbWindowsSizeEvent with the physical size of the screen
being expressed in the event. This physical size will then
be used to calculate a new dpi value.

This dpi-value-change will be used by the media query in a
CSS selector to change elements in the page. ***REMOVED*** will
specifically use this to dynamically optimize the size of the
fonts for a better user experience.

Implementation details:
Changes the width/height structure from a math::Size
object to a ViewportSize object which expresses width, height
plus diagonal_length.

Change-Id: I97affb732e1ad796d8be31d9a6e833e2a89dc3aa

Former-commit-id: 5386dccefec19dee1f51dc6cc3450615b653afca