Fixes memory reporting on nxswitch.
b/118403625

The problem was that the memory tracker would access thread
local storage to see wheather memory tracking was enabled. This
was used a thread guard to prevent reentrancy (a problem since
allocating memory will trigger memory reporting while reporting
memory and thus a stack overflow).

This is fixed by making an ad-hoc TLS system for the memory tracker
that guarantees memory allocated for the TLS will not go through
the standard memory reporting system.

Change-Id: Id997cfcc66090673c90622a1b1d4ab9093e5ac0a

Former-commit-id: 607b6c4bffc224259c5bcedd68403ae51f3cffc4