Filter fonts.xml to include only the configured Cobalt fonts.
Make checked-in fonts.xml the source of truth about what fonts are in
which package, rather than having a bunch of conditionals in the GYP
file. A script generates a filtered fonts.xml for the output content,
and provides a list of font files that need to be copied.

Don't copy any font files except Roboto-Regular-Subsetted.ttf for
Android since it falls back to system font files.

b/118150371

Change-Id: I4a3e2d7483db46f9440b0b4fa85937ad0646ce9c

Former-commit-id: 28e90f9ae988bb0785a49656d515d15057f5521d