Makes a build deploy error non-fatal.
For some reason the nxswitch build is failing because a directory
which the os has flagged as not existing, fails to be created.

Example:
http://cobalt-build.corp.google.com:8010/builders/nxswitch_debug/builds/425/steps/ninja%20all%20debug/logs/stdio

I am not able to reproduce this locally.

This cl will make that failure non-fatal and also print out
stats about the file for debugging from log files.

Change-Id: I2b2e0036c5a6a15791b73374fdc1838c9ed78c46

Former-commit-id: e14e470363dd293aed597460187063106ccdc2b6