Move logging backend out of DebugHub into LogComponent
Use the inspector protocol to communicate log messages to the debug
frontend rather than a custom callback in a custom IDL.

In order to avoid spamming a remote devtools session with a bunch of
Cobalt logs, use a custom "Log.browserEntryAdded" command that is
identical to "Log.entryAdded" but not recognized by Chrome devtools.
The overlay debug console does recognize it to show the logs.

b/31266139
b/118445468

Change-Id: Ib9b3a67dffccfb902be1109e88f12000c181af52

Former-commit-id: ffdbca6f5b811528476fc4eb048a64d922594db2