Fixes deadlock caused by a reentrant call to ~LogMessage.
This was causing win-win32 tests to deadlock occasionally
on buildbot.

Cause:

A global mutex in starboard::logging will be locked twice by the
same thread, which causes a deadlock.

starboard::logging calls starboard::net_log which will occasionally
produces an error. This error is logged which then tries to lock
the mutex which is already locked.

Here is the stack trace of the deadlock:

  SbMutexAcquire()
->starboard::LogMessage::~LogMessage()
  SbSocketBind()
  starboard::CreateListenSocket()
  starboard::NetLogServer::NetLogServer()
  starboard::NetLogServer::Instance()
  starboard::NetLogWrite(const char * data)
->starboard::LogMessage::~LogMessage()
  starboard::LocalizedStrings::LoadStrings()
  starboard::LocalizedStrings::LocalizedStrings()
  starboard::ApplicationWin32::ApplicationWin32()
  StarboardMain(int argc, char * * argv)

To fix this I added a reentrancy guard to the log_message.cc
so that if the ~LogMessage() is called twice in the same
callstack that the second call will return immediately.

I also moved ThreadLocalBoolean to starboard/common codebase
to prevent duplicate code.

Change-Id: I3ecf1b50d8d4748e37f6d3c75783e211fc34b060

Former-commit-id: eae7cfcc53cc8877dd5884a4157aeea883d4fae3