Copy ucrtbased.dll from windows sdk path rather than locally.
Previously the ucrtbased.dll was saved locally in the repostory.
This becomes a problem when upgrading the windows sdk since
there will be a version mismatch between the ucrtbased.dll in
the repository and the version expected to run.

This cl adds the ability for sdk_configuration.py to generate the
ucrtbased.dll from the sdk based on the _WIN_SDK_VERSION.

Removed in platforms xb1, mainappbeta, mainappfishfood, youtubekids
  Local copies of the ucrtbased.dll
  copy_runtime_dll.py

Change-Id: Ia6b2509e743afc918c7b4ff4657e3e44a0980ec6

Former-commit-id: f6318b25973d6d1fc071e4c44989003c862b0023