Fixes incompatibilities between ANGLE and msvc 14.15.26726
Starting with MSVC++ 14.15 std::future::wait() will throw an exception
with the message "Illegal to wait on a task in a Windows Runtime STA"
whenever it is invoked from the UI thread. This define signals a
workaround to avoid this exception by invoking future::wait() on a
separate thread and then immediately doing a thread join.

b/118398612

Change-Id: I6e16bd749261d8a441576ae52fbf35c673782e4c

Former-commit-id: b8061aa6c39633963ab42890b6d3821531271287