Implements platform specific sdk for windows platforms.
This change allows the multiple platforms that are compiled using
Microsofts sdks and compiler toolset to be versioned. Previously
all versions used the same global version of the sdk and msvc toolset.
This is a problem when upgrading since all platforms would be updated
to the same version.

It's important to note that this is a behavior-preserving refactor.
No platform will change their sdk versions as a result of this cl.

Change-Id: I4deee2137f157a7130dc53fd02f770eb15647d6e

Former-commit-id: 4d00b33ce60989cc8e92b352ec872a68c7496380