Enable build-time snapshot generation on all 19lts Android platforms
Build-time snapshot has been turned on for all v8 platforms on trunk
for more than a week now, let's turn it on for all 19lts Android platforms
to prepare for the upcoming new release!

Added a condition to enable some flags and defines only on target
platform, this is not needed on trunk.

b/111220585

Change-Id: I593219d15c700644aa59da6e720ba620c37cf19b

Former-commit-id: 23ac0eddcca7fe6b78cbfacee961a86ec23aa31e