Fixes issue with issue of new msvc compiler and internal linkage.
SupportsAddRefAndRelease<base::`anonymous-namespace'::NoRef> produces
an error with the new MSVC 14.15.26726 compiler.

The issue is the fact that the bind_unittest.cc is under an anonymous
namespace. This means that there is internal linkage and this apparently
causes a problem with a template instantiation

This is the specific error:
warning C5046: 'base::internal::SupportsAddRefAndRelease<
   base::`anonymous-namespace'::NoRef
>::BaseMixin::AddRef': Symbol involving type with
                       internal linkage not defined

The fix is to move everything in the unit test into a non-anonymous
namespace, which allows non-internal linkage.

Change-Id: I2f530507b103f20469bbdac8598e8ce4e8900fce

Former-commit-id: 9aadcfc40b67da159f63b8dea23d076703c397ef