Re-enable performance/tracing panel
Enable the "profiler" module, as well as its dependencies in the
devtools_app, worker_app, and shell.

Enable 'timelineTracingJSProfile' experiment by default so that the
necessary "v8.cpu_profiler" category is enabled.

Disable the HeapProfilerPanel in the profiler module, leaving what's
necessary for CPU profiling. (Heap profiling and CPU profiling are
interwrined in the profiler module.)

Disable screenshot and memory checkboxes, and settings button in
performance panel toolbar. Also use 'false' as the default setting for
showing screenshots.

Change-Id: I8a8f22f60f8cd3c6b2c302b6d0a1009f2269f9a0

Former-commit-id: 03dae3d630c29d6dd1076bc18dd9f84c796750e1