Adds error output when windows build tool fails.
Example:
http://cobalt-build.corp.google.com:8010/builders/xb1_gold/builds/3592/steps/ninja%20cobalt_deploy%20gold/logs/stdio

In this example the command failure occures with:

FAILED: C:\buildbot_win\buildbot_install\env\python.exe
gyp-win-tool action-wrapper environment.x64 ...

This command runs one of many internal commands. This internal
invocation will swallow all stdout and stderr output.

This cl changes the behavior by checking the return code of the
internally run command, and if it is non-zero then the command
+ args + stdout + stderr + error_code is printed.

If the command fails in the future (it's a rare occurence) then the
developer can be clued into the cause of the problem.

Change-Id: Iadb5fc611c2c8f537d44af39cbadcac4bec1434b

Former-commit-id: 30b538e780e4982f52582005e13b420bb04bc16d