Adds the device_server implementation.
This allows a device server to act as a proxy for the
nxswitch device. Nxswitch builders can connect to the
device server and make a run request.

This request will include a bundle.zip file which contains
the binary files for nxswitch + an info json file that will
contain a manifest of what to run.

Change-Id: Ied29ece63b702d1b14c78ce43eb8fa9f655fcec1

Former-commit-id: b9327c45be6db25de6d3d6003a6e2c5564f06b7a