Hook up TracingCpuProfiler to isolate.
This is the lynchpin that allows the debugger to enable the profiler
since the |TracingCpuProfiler| hooks itself up as an observer so that
its |OnTraceEnabled| function is called by the |TracingController| when
the Cobalt |ScriptDebugger| calls |StartTracing|.

When tracing is enabled with the "v8.cpu_profiler" category, the
|TracingCpuProfiler| instantiates a new |CpuProfiler| and starts it on
the isolate's main thread. The |CpuProfiler| starts a processor on a new
thread to regularly have the platform-specific sampler provide the
register state to be recorded and processed.

This is taken directly from V8 in Chromium.

b/27268575

Change-Id: Ia1bb646c25f96e7a0453c6c6d8221d04f90bbdf8

Former-commit-id: 33fba1c7caf7b9866a64dbbb8f371b37d3a866fe