Refactor Loader to fully encapsulate Decoders.
Refactor Loader to fully encapsulate Decoders such that it rather than
the Loader Factory is responsible for the creation of Decoders. This
allows the Decoder to have a callback to the Loader so that the Loader
can be notified of the Decoder's status. This is important as it allows
the Loader to better handle suspend and resume states, avoiding the case
where Loader Start() can be called twice despite the load having been
completed.

b/119536712

Change-Id: Ic347bdc2d4dbc7a05a6df331b45424680ac1bde2

Former-commit-id: bd3ab4e791df1b893753b933e742cc4f4307950f