Use our own V8 TracingController rather than the default.
The |V8cTracingController| is a wrapper that calls through to an
instance of the default V8 platform |TracingController|, but provides
its own |TraceWriter| to pass the events to a |TraceDelegate| rather
than just writing them to the log.

Its internal |TraceWriter| leverages the V8 |JSONTraceWriter| to convert
the trace events to JSON so that the |TraceDelegate| gets events already
in the JSON trace event format understood by the devtools frontend.
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU

In addition, the |V8cTracingController| accepts a list of categories
when tracing is started, which it adds to the |TraceConfig| that is
passed to the the default |TracingController|.

Also changed to use |NewDefaultPlatform| instead of the deprecated
|CreateDefaultPlatform|.

The |ScriptDebugger| is the surface through which this is all used,
declaring both the |TraceDelegate| interface mentioned above, as well as
|StartTracing| and |StopTracing| functions to control when tracing
occurs. Those have been implemented in both the |V8cScriptDebugger| and
the |StubScriptDebugger|.

b/27268575

Change-Id: I7388421d2c381784ab3f3fdc1d3db4596d98ef4e

Former-commit-id: 06b49cb8659872ad49a820a90ee5e91454f96882