Adds win_symlink_fast.py
This adds accelerated versions of reading reparse links. This
is necessary for device_server MakeBundle, which will test input
files on whether they are symbolic links. The previous technqiue
of using the shell commands to read symbolic links would take
approximatly 48 milliseconds, while this new version speeds this
up to be < .1 milliseconds.

Given that windows functions for reparse point manipulation tend to
raise security exceptions when not running in non-admin mode for some
versions of windows, the top level ReadReparsePoint() function in
win_symlink will attempt to run the fast versions and on exception,
will fallback to the shell command versions and print the exception.

Change-Id: Ib406d5c01125070823176706645da1a48024fa95

Former-commit-id: a87b92fd124f0301185572fdaa7492f920a48c72