Allow limiting WV session updates via command line
Add command line parameter "--maximum_drm_session_updates" to non-gold
version.  It will be picked up by our WV based SbDrm implementation to
limit the number of session updates SbDrm will send.  Once the
limitation is reached, the session update request will be failed with
kSbDrmStatusQuotaExceededError, this allows us to test the behavior of
YT player when the session limit is reached.

This is only implemented for SB_API_VERSION 10 or later as otherwise
kSbDrmStatusQuotaExceededError is not supported.

b/123548830

Change-Id: Id5df44ee5138cbe21c2792f7cbc87da5536f514c

Former-commit-id: 42efa60ed5e705d02a5fdf1cb1d11eeda66a781d