Remove redundant command and event string constants and literals.
Make the |CommandMap| helper class optionally able to prefix mapped
command names with a domain, allowing the agents to re-use their
|kInspectorDomain| constant rather than having to repeat it when setting
up the commands in the map.

Inline string literals for command/event parameter field names and
values so that it's more apparent just what's in the JSON messages at
the point in the code where they're being assembled.

Change-Id: I8fae0d9c071ed7d9b5daf0846f768b087ebdca85

Former-commit-id: b7bccd857bb192f444bc0e4dbfd0ab2d7221150f