Add unit tests for webaudio resampling
Check that the resampling logic is working correctly
when resampling to both higher and lower sample rates,
and when resampling webaudio of types float32 and int16.
Check that the expected resampled output and the actual
resampled output for a simple wave function are within
a small margin of error of one another.

b/116029316

Change-Id: I350a90ffde265c7eea34546d3941c6a0e1ca5fec

Former-commit-id: 5642bfa4a807011c035e5a20ca55889810599f81