Adds timeout to NetLog/NetArgs.
If xb1 is installed via the starboard launcher then cobalt
will wait for a connection to net args and then also to net log.
This presents a problem if the cobalt binary is run again
from the native launcher on the xb1 device, since there will
be no client to connect to transfer net-args/net-log and the
binary would wait for ever. Now the cobalt client will wait
upto a timeout period and then continue launching if no client
will connect.

Change-Id: I11b07bd7049916a6d1f1ed1619cd7863405c443d

Former-commit-id: 6f0170414388906ac502d00120fc06503d76d3ad