Flush after SbStorageWriteRecord temp file rename() occurs.
On some platforms it was found that the rename() would frequently
be delayed itself and so was susceptible to power outages. It
was found that calling SbFileFlush() on the temp file (whose
file pointer remains valid through a rename) would solve the
problem.

This was reported by multiple partners including Samsung and
RT-RK.

Contributing CL:
  https://cobalt-review.googlesource.com/c/cobalt/+/5270

Change-Id: If0668169262736f3e944008fd04887bd97db192f

Former-commit-id: b32092d957dcfd24387b2cb7903dff06e65734dd