Fixes FAST symlinks for Windows Server 2016 when not supported.
The fast implementation of creating windows symlinks
through the win-c-api bridge only works under certain conditions.
For some unknown reason, these links will fail to be created on
Windows Server 2016.

Therefore, additional logic will be used to fallback to the slower
shell call to do the equivalent operation.

This should fix the nxswitch build, which was broken because of
this issue.

Also sprinkled in some universal_newlines to fix more of the double
spacing problem in the buildbot web log.

Change-Id: I7acc4a00d6d160a58a273d1e8ab3c1664aca0eac

Former-commit-id: 1974e52c0d0cdf9166bbef19905d8aa517021f11