Restore absolute paths for link path in win_symlink
This will help the ps4 build, which is using this functionality
but is inputting a relative path and relying on the win_symlink
to convert it to an absolute path. This feature was removed when
I added relative symlinking and the result was that the ps4 build
broke.

In all cases, the link_path can certainly be unconditionally
converted into an absolute path.

This code path was tested through the device_server bundler and
also the nxswitch device_server test.

Change-Id: Ia0de673875ed7834420c560649d9207a878fb196

Former-commit-id: c51bca0312c39c82607631bae00ad35db61071d0