Initial implementation of a basic CSS agent.
This is just a start and still needs work!

Only CSS.getMatchedStyledForNode() is implemented, and it's incomplete.
Only the rules directly matching the element are returned, without the
inline style nor inherited rules.

It's quite slow and has lots of room to be optimized. The command is
implemented completely in JavaScript, and even that implementation is
inefficiently looping over all selectors in the document to search for
which ones match the element. And, Element.matches() is a polyfill.

b/112559341

Change-Id: Ic42d2b7c1eafb1ca9180a6b98b21ebfcddcea2dd

Former-commit-id: 552ec481d47694048406858baa6d0ccfc81ca0a1