Remove animations from all AnimationSets when they are destroyed.
web_animations::Animation objects usually register themselves
explicitly with all AnimationSets they will appear within,
indirectly through registering with their target Elements and
their document timeline objects.  Things must be done this way
to conform to the web animations web specification.

However there is a case where they might also be added to
AnimationSets that are not part of the web specification.  In
these cases, when the Animation is destroyed, we must make
sure that the Animation is removed from all AnimationSets, even
the ones the Animation did not explicitly register itself with.

This CL makes Animation objects track the AnimationSets they
belong to, and explicitly removes itself from any AnimationSets
that it is still in when the Animation is destroyed.

b/123545481

Change-Id: If04be79da4ca54122d045f5cd359e19ffe3c546f

Former-commit-id: 68597967fb3619a5283c2f5070509c006b23d20b