Handle locale change notification
This change implements a class that listens for changes to the system's
locale settings. If a change is detected while the application is
stopped, an instance of this class will cause the application to exit.
On subsequent launches of the application the most recent locale
settings will be adopted.

b/118687968

Change-Id: I75d49ff99b88b4bb25ee77105c06831e7be0ad9a
(cherry picked from commit ac3e7858c3620b5a0751bc378dae01fb610ab3f0)

Former-commit-id: 35a392d85c95af0709f5464c03bc69cfd98f6b38