Handle locale change notification
This change implements a class that listens for changes to the system's
locale settings. If a change is detected while the application is
stopped, an instance of this class will cause the application to exit.
On subsequent launches of the application the most recent locale
settings will be adopted.

b/118687968

Change-Id: I75d49ff99b88b4bb25ee77105c06831e7be0ad9a

Former-commit-id: cc8fdcf57dd3d976fb6fa373510b004bb0d61a40