Add an option to set test server binding address.
Add an option (--server_binding_address) to set the ip address
that should be used to create the test server used by the black
box tests (default was to use the IP address of the hostname,
however for machines having more than one interface, such as
Mobile Harness servers, this would not work).
Simplify ThreadedWebServer's __init__ function.
Pick a random port for the server instead of first trying 8000.

Change-Id: I86797f3fe6866f37089e2f995961b0f99ce7e63d

Former-commit-id: 7c61874977f56813988d89b653689dc8383ad6ab