Fix issue with XHR loading updates not having latest data.
After this change:
https://lbshell-internal-review.git.corp.google.com/c/cobalt/+/114842/4/xhr/xml_http_request.cc#712
,

we were firing the readystatechange event *before* we had appended
the new XHR data, so the latest data was not visible to JavaScript
when JavaScript was told that new data is available.

This is probably not good all over the place, but in particular,
it was causing an issue with ***REMOVED*** events lagging behind the user's
input by one event.

b/123559324

Change-Id: I2e0acf7c55809d6817bd3c691f9d7c3a568b4171

Former-commit-id: 0bd747d400a378ec24236620115d0fc824ab19c1