Allow debugging through navigation.
Save the debugger state before tearing down the old WebModule and pass
it back in when creating the new WebModule.

The lifecycle of the DebugModule, DebugDispatcher, and all the agents
is still tied to the WebModule, but now they are "frozen" and their
state is saved just before navigating so the new instances can be
"thawed" to restore the saved state.

When exiting Cobalt (rather than navigating), the DebugModule dtor
freezes itself without actually saving the state in order to give
everything a consistent mechanism to clean up.

The DebugDispatcher now has a movable helper ClientsSet class that
transfers the set of attached clients through the saved state to the new
DebugDispatcher instance after navigating.

The agents now register/unregister their domains with DebugDispatcher
in their Thaw/Freeze functions rather than in the ctor/dtor so that
hybrid JS agents don't have to wait for garbage collection to release
the wrapper's ref-counted reference. (All agents have the Thaw/Freeze
functions to do this registration, but at this point only the logging
agent and script-debugger agent have any meaningful state to save.)

The V8cScriptDebugger no longer keeps a single long-lived inspector
session. Instead, sessions are connected/released when thawing/freezing
the ScriptDebuggerAgent. This allows us to save the V8 inspector state
as part of our overall debugger state so we can restore it in a new
session after navigating. Changing sessions makes the V8 inspector send
all the necessary events so the DevTools frontend recognizes the change
of context associated with the navigation.

b/129301981

Change-Id: I47fb02138a5b4b1aeb51f7f85588e770a559fdd0

Former-commit-id: c188a87fff30903c31e4030bc5409d9a0a219165