Make decode target content_region be specified in floats, not ints.
This permits platforms to specify texture width and height as
(1, 1) and then use normalized coordinates when specifying the
content region.  This is useful to do on platforms where you don't
know the exact texture width/height, such as on Android TV where
we may not know the exact texture width/height of decoded video
frames (the platform may be backing the decoded frame with a
texture larger than the video frame size).

b/111242817

Change-Id: Ib90b90a0cf9073dbf0c2ad2df39a81f4fef3138e

Former-commit-id: 27edac94194a47604e91bab154c49b7b8eb005b2