Remove leaking code from libcrypto.a with preprocessor macros
This change wraps #ifdef's around function calls that are no longer
needed with a change to the BoringSSL ninja configuration that removes
certain files with API leaks.

The functions that are removed were implemented in files that are no
longer being included since they were dead code that introduced API
leaks (lbshell-internal-review/128904).

b/128434484

Change-Id: Iae12bdc2083ea5ccd99bccab90d41ea45c36bea1

Former-commit-id: a34a24222fc92fc65863f09cc0395b65cb4fce23