Protocol error instead of DCHECK when agent script doesn't load.
Load the agent JavaScript in Thaw(), and remember whether or not it was
successful. If the JavaScript failed to load, return an error when
DevTools tries to enable the agent's domain.

This fixes layout_tests, etc. which don't include the debugger backend
content, and therefore can't load the debugger agent's JavaScript code.

Change-Id: I8d2eb23893bec46c048ea53fbe2e206b1d733c7e

Former-commit-id: f0e746ad1d9df151c3d9fb02cf49ebc7246f83d5