Fix sporadic Linux crashes related to GoogleSpeechService.
On linux-x64x11 when running "--input_fuzzer", there would be an
occasional crash with a callstack showing GoogleSpeechService.

This happens when GoogleSpeechService is started on a platform
that doesn't have an API key (e.g. linux-x64x11).

On platforms that don't have a speech API key, we shouldn't even
try to use GoogleSpeechService, so we now simply generate an
error immediately in that case and that's that.

b/126447321

Change-Id: I6ace508fb997fc0b7519459e54247d314800956e
(cherry picked from commit d89b94855337d2a9e96ea2b3acc5ac920df3745c)

Former-commit-id: a709d576270a5c02f78a2b62f3e9db5c46641ac2