Fix sporadic Linux crashes related to GoogleSpeechService.
On linux-x64x11 when running "--input_fuzzer", there would be an
occasional crash with a callstack showing GoogleSpeechService.

This happens when GoogleSpeechService is started on a platform
that doesn't have an API key (e.g. linux-x64x11).

On platforms that don't have a speech API key, we shouldn't even
try to use GoogleSpeechService, so we now simply generate an
error immediately in that case and that's that.

b/126447321

Change-Id: I6ace508fb997fc0b7519459e54247d314800956e

Former-commit-id: ca1062b01d8c2ee105470a8d1838162b1c7990c3