Move log synchronization mechanism behind Starboard
Log synchronization has been moved behind Starboard. The Starboard
extended API refactoring causes the mutex and minimum logging level used
for Starboard logging to no longer be guaranteed to be consistent with
Cobalt. To combat this, log synchronization has been moved behind
Starboard and a requirement that logging methods, such as SbLog or
SbLogRaw, be thread-safe because they will likely be called from
multiple threads.

The minimum logging level is no longer set by Cobalt, and is instead set
by grabbing the value as a command-line argument.

Additionally, Android previously relied on SbLog to pause for a
millisecond so that the reading process can handle the logs without
dropping any
(https://lbshell-internal-review.git.corp.google.com/c/starboard/+/96052).
To implement our change for SbLog and not SbLogRaw, some of the code
that was implemented in //android/shared/log.cc has been moved to
//android/shared/log_internal.cc to be used by SbLogRaw.

b/131339507

Change-Id: If6f4e6d1584bc59e0e54eed92ad55560b42ad33f

Former-commit-id: 2e6854fe53301bdc1b57bef6b4ee0b78da988903