CSSOM Fix blockification.
This fixes CSSOM blockification in a few ways.

Blockification should change the computed style, not the initial style.

Flex items are blockified, which requires access to the parent style to
determine if the parent is a flex container box.

When an element is 'inline-flex', it should be blockified to 'flex'.

When an element is blockified, that should
not affect the static position that is (in some cases) used for absolutely
positioned boxes.

For example, this fixes the bug where absolutely positioned boxes without a
coordinate are placed at the beginning of the line instead of at the in-flow
line position. A layout test for this is added.

Since blockification is applied during computed style calculation, it may
not be applied if the computed style is modified afterwards.  Since the
computed style declaration should be immutable, this removes the setters
from CSSComputedStyleData and moves them into a derived class
MutableCSSComputedStyleData to make it very explicit which code can modify
it.

This also adds computed style unit tests for the blockification.

This includes a fix for b/29370980

Change-Id: I2c491cbbbc6086cf5509dae36daa3c68f902ebc3

Former-commit-id: c435a96a6d87e9d623a3f08b664c43915826d0ad