Add tests for SbMediaSetAudioWriteDuration
Add tests for SbMediaSetAudioWriteDuration in player_filter_tests:
Test that we can successfully decode an audio sample that's only 0.5
seconds long. Test that we can decode to the end of a longer audio
sample, while only providing 0.5 seconds of data at a time.

Filter on Android because it already fails similar AudioDecoderTests
cases like MultipleInputs, and so we don't expect it to pass the new
tests either.

b/123957391

Change-Id: I1712ea84aae13ab5fef3c7a56750002aed46b160

Former-commit-id: 018a57460685d42cc84abce3fb7668ad1f31b2b5