Default to not using Starboard for image decoding on Linux.
While we still keep the option of enabling this feature via
a command line flag, we default to false instead of true now.

This helps with a problem where on Forge enabling Starboard
image decoding can result in a number of screenshot tests
failing due to 2s timeouts waiting for image decodes to finish.

b/131715429
b/122362330

Change-Id: Iea3eaff805e74cbb85a52461325fce4cec2f97b5

Former-commit-id: 2ab82174608ecf621dcb0b48ba5b5f1ba9f77a77