Don't fire timeupdate on initial preroll
When the media pipeline is started, it notifies the web media player
that the initial preroll is finished, as if there was a seek.  However,
the behavior of the initial preroll is slightly different than a normal,
explicit seek, as it shouldn't cause the fire of a timeupdate event.

This change adds an extra boolean parameter to seek callback so the web
media player can know that the seek callback is called due to an initial
preroll or not, and to decide whether a timeupdate event should be fired.

b/132156457

Change-Id: If813c84a21080834a22b3e677fce2ffbab2427cc

Former-commit-id: d02db45c7de77d951150b57f462cf76109b47238