Send content only for the target being run to raspi.
Rather than rsync-ing the whole (filtered) out/ directory to the device,
make the source of the sync just the deploy directory of the target
being launched.

The destination directory on the device is the same for all targets (the
basename of the out directory) so that when several targets are run one
after another on the same device, the content that was already sync'd
doesn't have to get sent again.

b/133264398

Change-Id: Ie46c40375c8bc389e54ab10ad8998fb9c543c58f

Former-commit-id: 69afd40f6c78ab75fb0db44d8c822ebe6e07aedb