Move call to GetHostCompilerEnvironment() outside of __init__
gyp_utils.GetHostCompilerEnvironment() gets called in the __init__
function of gyp_configuration.py for the Android platform.
When running the performance test from the code copied on x20
this makes the test fail as none of the toolchains and
toolchain-related scripts are available on x20.
The gyp_configuration.py for other platforms callis this method in
GetEnvironmentVariables, however for Android the resulting
environment variables are only used in GetHostToolchain instead,
which is why this change does not move it to
GetEnvironmentVariables, but to GetHostToolchain instead.

Change-Id: Ife1f90058f7b4db14b8c20822b715934e1bdf989

Former-commit-id: 17644cf4ccba40f8b2c68c570c2f68422a076166