Switch Android to use the LLVM "lld" linker.
Due to a bug in the Android NDK, the arm64 linker was always different
than the other architectures ("bfd" rather than "gold"), which could
account for why it was the one getting killed when it ran out of memory.

In the latest NDK 19, they have now introduced the "lld" linker and
encourage testing apps with it, so here we go.

Also removed the use of the "--no-keep-memory" flag that was an
ineffective attempt to reduce memory.

As an added bonus, the linker is so much faster! For example, the time
to link libcobalt.so with each linker (on my Z840 workstation):
  bfd : 50s
  gold: 21s
  lld :  4s

b/64942033

Change-Id: Iee245d33ce5204eb89830be0636bda87b465bfda

Former-commit-id: 85fc8193e75ac4ffdb79a842f2ee865005875bb2