Implement *fake* hdr support for av1/hevc
Claim support of hdr 10 bits video for av1 and hevc, decode them in 10
bits and render them in 8 bits, so we can at least test the interfaces.

Note that the vp9 10 bits decoding interface is ready, but it is not
supported by the current libvpx on gLinux.  So the support of vp9 10bits
is still rejected in SbMediaIsVideoSupported().

Change-Id: I5d8fc4539ba0cfae613cb1bf8f66421e8014de34

Former-commit-id: 1116f4d2022d5f8805c84216836fdbe85c177774