CobaltArchive: Default include paths are improved.
If a target platform does not supply a GetDeployPaths() in
it's gyp_configuration then the paths included by default are
now good enough for platforms that don't have a deploy directory
structure such as win-win32 and linux.

This CL unlocks the win-win32 archive to be built and run
by the cobalt archive runner.

Change-Id: Idb7fbb3ed18c7fc399859d8daf984ce0d2b8cbae

Former-commit-id: c29b12cc30b7354256538352725dd210768ef60e