Download the specific Android NDK that we want.
Don't use the sdkmanager to download the NDK as there is no way to
specify the version.

We want to use a specific NDK until such time as we're ready to upgrade
so that builds are repeatable as we go through a release cycle.

Also, each and every upgrade between NDK 14 and NDK 20 has broken the
build in some way, so we can no longer let that happen at the whim of
the NDK release cycle, and need to regain control of when we upgrade to
avoid fixing things as an emergency reaction.

b/64853164
b/134680168

Change-Id: I6684034b31a84dba4d2bae92dcd61bed67c07ee0

Former-commit-id: 5da023d90b7248529f094d8d1c87221fbb8577fd