Don't unzip left-over bundles on the raspi
We previously zipped up the content directory into a "bundle" as part of
the build, and that zip would be copied to X20 and then rsync-ed to the
raspi. The launcher was responsible for unzipping it on the device
before running the executable.

Now that we no longer create that bundle, we also don't want to unzip
any random zip files that may remain on the device since that takes time
and clobbers the updated content files that were just rsync-ed to the
device.

b/133264398

Change-Id: I2098b98b9b584e1af6f725d4fd24dfb396ce5875

Former-commit-id: 816df2e4d2a9b6ec1a24154fb68638c50a6db085