Avoid accessing task_runner on threads without MessageLoop
When testing if the current running thread is the wanted thread, our
current code sometimes try to access the task runner of the current
thread which is not guaranteed to exist if the thread does not have
a MessageLoop. This CL should change all such code occurances in Cobalt
code that might result in a crash.

b/134682690

Change-Id: I4b049af5d603d4abb41c8f04a2c2e3b5541c7e52

Former-commit-id: a7260e2bc74b56f3cd4c92294a1ab18a9b4e5046