Fixes filelist.py for rel symlinked files on linux
b/134977445

Android uses relative symlinking. When I attempted to create
an archive build for the platform then these relative symlinks
were hit and this caused a build failure.

In the future we will only add the apk for the android archive
build without any symlinks. However, this use case allows me to
fix the filelist / archive creating for linux symlinks in general.

Change-Id: I6fd5b26daa5b6ad223af51673555b9b62857fc32

Former-commit-id: 86c1fd384a72ddc72b0294e98c9c3039e5e71875