Refactor ResourceCache
Abstract non-template functionalities of CachedResource and
ResourceCache into non-template base class CachedResourceBase and
ResourceCacheBase.
There is no change of functionality, the implementation of quite a few
functions are left in their original place so it is easy to review the
code.  They will be moved into .cc file in a follow up CL.

b/119883416

Change-Id: Ib05ce583602f64657c3ecb8c1219c851e9ad1d72

Former-commit-id: 8084032f82f5a53d3e84268f057d9642ac445a30