Reduce error spam from: https://lbshell-internal-review.git.corp.google.com/c/starboard/+/135700
Example:
[7968/7988] ACTION Collect content: media_session_test
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.
You do not have sufficient privilege to perform this operation.

http://cobalt-build.corp.google.com:8010/builders/ps4_tests/builds/5555/steps/build%20Unit%20Tests%3A%20devel/logs/stdio

Error message is coming from mklink /d, which is expected to
maybe fail on older versions of windows (and has a fallback of
mklink /j)

Change-Id: I10133dce35ea3c27fddbce10dd0cb6622de07a1c

Former-commit-id: 1c778ea223ecbdada5e173459d7f721e0ea0c574