Port network-thread priority performance fix
It was found that having the network-thread run on low priority improved
the performance of playing videos on certain android devices. The
default thread priority was set to LOW for that release, but now we want
to redo this change in favor of a more portable implementation.

Further testing on high-end devices such as NVIDIA shield and PS4 showed
that the performance was not negatively impacted by setting the network
thread priority to LOW, so this change is a performance gain for low-end
devices without impacting the high-end.

b/132704285

Change-Id: I68903ea1e79e7fa1f7b82c0daf6049cced0c005c

Former-commit-id: c514f29c64db547341ee12b15278f1de48b231f6