Add Box::GetMarginBoxTransformFromContainingBlock
This also refactors GetTransformedBorderBoxFromRoot() and
GetTransformedBorderBoxFromContainingBlock() to use the
new function so it will be tested by all the unit tests
for those functions.

Also add a new unit test for getBoundingClientRect which
uses GetTransformedBorderBoxFromRoot() -- this tests
evaluation of nested transforms.

The new function will be used by hybrid nav to support
transformed navigation items.

b/129365262

Change-Id: I41a1ef43f95318ac9ea5335a068cae6c31ba9c57

Former-commit-id: dc349ac814e0194648613adc284922533475939e