Flexbox: Add FlexItem
This adds a FlexItem class that allows using main and cross axis properties,
sizes, and positions, redirecting it to with and height as needed depending
on whether the flex container is row or column.

This makes column flex containers layout better.  There are still column
layout bugs, so layout tests are not yet included.  They will be added with
the bugfixes.

b/27337434 b/137045624

Change-Id: I58e31fd60f54b9cc7b17afdecb546dd2c9b32025

Former-commit-id: 9a6e0812e24d03c0d38926ad5ccbfbb4f5d1af39