Generate the build number in gyp_runner, not cobalt_configuration.
Save the build number in a "build.id" file in the out directory from the
gyp_runner, then use that as the source of truth when configuring the
cobalt build. This makes populating the cobalt application variables
work without running gclient or contacting the build ID server, enabling
it to run on mobile harness where those aren't available.

Restrict usage of gyp_utils by replacing the remaining two usages of
gyp_utils.GetHostCompilerEnvironment() with the starboard.tools version.

b/136217063

Change-Id: I247b37b4ee488c441a039cbaf8e2996dd07de554

Former-commit-id: 07869a2ccda382563816ecb4ddfa8d198ffe99e1