Make Get[Environment]Variables() run without side-effects.
Generate the build number in gyp_runner, not cobalt_configuration, and
save it in a "build.id" file in the out directory. Use that as the
source of truth when getting Cobalt variables rather than running
gclient or contacting the build ID server, which aren't available when
running black box tests on mobile harness.

Install SDKs, etc. in a dedicated SetupPlatformTools() function rather
than as a side-effect of GetEnvironmentVariables(). This is also the
place to raise a RuntimeError if something will prevent building, rather
than failing while getting variables.

Restrict usage of gyp_utils by replacing the remaining two usages of
gyp_utils.GetHostCompilerEnvironment() with the starboard.tools version,
and passing the build number into Android's SetupPlatformTools().

b/136217063

Change-Id: I09e394736f83b0862e167f0dbd2cb5a7987e7c44

Former-commit-id: e03dbf37af38d612c4f9c2d62d2fcbac2e766a45