Retry flaky test case failures, reporting cleanup
This change makes it so that flaky tests are re-run on failure. If they
succeed within the retry limit they are marked as passing. Additionally,
this change includes various output formatting changes.

This change makes a few improvements on the test_runner.py script.

1. Flaky tests are re-run on failure. If they succeed within the retry
limit they are marked as passing.
2. The reporting has been cleaned up and now reports additional
information, in a cleaner fashion.
3. Minor other formatting and style changes.

b/36592961
b/132435128
b/137944372

Change-Id: Iae69f642d71356e13c6802c59a5ea44ba3ec621f

Former-commit-id: d7d0563f4c2a65c09101a21e46531c7a61127beb