Set an automatic BUILD_NUMBER gyp variable.
Rather than writing the build number to a "build.id" file in the out
directory, just make it available as an automatic gyp variable.

The source of truth for BUILD_NUMBER is |gyp_utils.getBuildNumber()|,
called when running gyp_cobalt. For OSS Cobalt that reads it from
|cobalt/build/build.id|, which not present in the internal repo. For the
internal repo, it gets it from the "carbon-airlock" version server based
on the hashes returned by "gclient revinfo -a".

b/136217063

Change-Id: Ib315179275fde186861992369989285ffd864069

Former-commit-id: f214be95e70f5e46023923dc38c60fcda0ed2e3a