Output more info when all tests filtered, clean up output
When a test suite is run and all of the tests have been filtered out, if
there is a non-zero return code it will not output any information other
than:

Results not available.  Did the test crash?

We should only ever output this when there is a non-zero return code AND
the number of tests run is 0 AND the number of filtered tests is 0;

Additionally, this change moves the "TEST RUN COMPLETE" log outside of
the primary loop (spammy).

b/137957573

Change-Id: Id884351dfcef3395bd89539b2d94d353fb8a1eaa
(cherry picked from commit edb5071da205944f9865878e8df45818e942ad62)

Former-commit-id: 658e1e8b5acc98ff9d65c9b3b070e7f2a832c4e5