Output more info when all tests filtered, clean up output
When a test suite is run and all of the tests have been filtered out, if
there is a non-zero return code it will not output any information other
than:

Results not available.  Did the test crash?

We should only ever output this when there is a non-zero return code AND
the number of tests run is 0 AND the number of filtered tests is 0;

Additionally, this change moves the "TEST RUN COMPLETE" log outside of
the primary loop (spammy).

b/137957573

Change-Id: Id884351dfcef3395bd89539b2d94d353fb8a1eaa

Former-commit-id: 509dc4898b64e22809585797f14f29304adf853e