Flexbox content based minimum size.
This adds support for Flexbox content bases minimum size rules.

This also adds corresponding support for the 'auto' keyword for min-width
and min-height, and changed the initial value from '0px' in CSS2.1 to 'auto'
in CSS3 sizing: https://www.w3.org/TR/css-sizing-3/#min-size-properties

This moves DetermineFlexBaseSize() to FlexItem, moves the hypothetical main
size determination into FlexItem::DetermineHypotheticalMainSize, and moved
FlexItem creation from FlexLine to FlexContainerBox.

This corrects handling of flex item shrinking, and justify-content of
space-between and space-around with negative free space, and a corresponding
test.

Some tests are added to the acid pages for issues fixed in this CL.

This fixes b/137045763
This fixes b/137405652
This fixes b/137214916

Change-Id: Iae9205076062919a72de08007453c0f73eb4ff0b

Former-commit-id: 9236df3090b7bb5132ad466a5e288991de06644d